Add thefuck

+++ setup.sh
@@ -8,7 +8,8 @@ echo "Installing system dependencies"
 sudo apt install \
     dunst \
     entr \
-    pip \
+    python3 \
+    python3-pip \
     rofi \
     zsh
 
@@ -17,8 +18,25 @@ echo "Installing python dependencies"
 pip install pipsi
 
 
+function pipsii() {
+    python3=$(which python3)
+    for var in "$@"; do
+        pipsi install --python "$python3" "$var"
+    done
+}
+
+pipsii \
+    thefuck
+
+function cargoi() {
+    for var in "$@"; do
+        cargo install --list | grep "$var" || cargo install "$var"
+    done
+}
+
+
 echo "Installing rust dependencies"
-cargo install --force \
+cargoi \
     fd-find \
     lsd \
     ripgrep
@@ -43,14 +61,17 @@ function backup() {
 function bnl() {
 	dst="$HOME/$1"
 	src="$SCRIPTPATH/$1"
-	if [ -e "$1" ]; then 
-		backup "$dst"
-	else
-		rm "$dst"
+	if [ -e "$SCRIPTPATH/$1" ]; then
+        if [ -L "$dst" -a $(readlink -f "$dst") = "$src" ]; then
+            echo "Skip replacement of $dst, it is already the right target"
+        else
+            echo "Backup $1 to $dst"
+		    backup "$dst"
+	        echo "Linking $src to $dst"
+	        ln -s "$src" "$dst"
+        fi
 	fi
 
-	echo "Linking $src to $dst"
-	ln -s "$src" "$dst"
 }
